Time: 1301 ms (33.33%), Space: 59.1 MB (100.00%) - LeetHub

+++ 2179-count-good-triplets-in-an-array/2179-count-good-triplets-in-an-array.java
@@ -25,7 +25,7 @@ class Solution {
         }
         
         for(int i=0; i < nums1.length; i++) {
-            res += (long)left[i] * right[i];
+            res += (long)left[i] * (long)right[i];
         }
         
         return res;
